refactor: shared/ui select 분리
