+++ src/shared/ui/select/index.ts
@@ -1,0 +1,4 @@
+export * from "./Select"
+export * from "./SelectContent"
+export * from "./SelectItem"
+export * from "./SelectTrigger"
